test: drastically reduce code duplication in tests of models
